+++ src/index.js
@@ -1,7 +1,6 @@
 
 import React from "react";
 import ReactDOM from "react-dom";
-import Header from "./header";
 import Frontpage from "./frontpage";
 import './css/App.css';
 
@@ -9,7 +8,6 @@ import './css/App.css';
 export default function App() {
   
     return (<article>
-        <Header />
     <Frontpage />
       </article>)
 }
